Fixed dynamic filenames are not parsed correctly.
